[frontend] split entity settings visibility and references
