+++ opencti-platform/opencti-front/src/private/components/settings/sub_types/SubType.tsx
@@ -92,9 +92,8 @@ const SubTypeComponent: React.FC<SubTypeProps> = ({ queryRef }) => {
 
       {
         !isDraftWorkspaceType && (
-          <Card title={t_i18n('Configuration')}>
-            <EntitySettingSettings entitySettingsData={subType.settings} />
-          </Card>
+
+          <EntitySettingSettings entitySettingsData={subType.settings} />
         )
       }
