+++ opencti-platform/opencti-front/src/private/components/settings/sub_types/entity_setting/EntitySettingSettings.tsx
@@ -1,20 +1,13 @@
-import { Box, Tooltip } from '@mui/material';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import FormGroup from '@mui/material/FormGroup';
 import Grid from '@mui/material/Grid';
-import Switch from '@mui/material/Switch';
-import { InformationOutline } from 'mdi-material-ui';
 import { graphql, useFragment } from 'react-relay';
-import Label from '../../../../../components/common/label/Label';
+import Card from '@common/card/Card';
 import ErrorNotFound from '../../../../../components/ErrorNotFound';
-import { useFormatter } from '../../../../../components/i18n';
 import useApiMutation from '../../../../../utils/hooks/useApiMutation';
-import { SETTINGS_SETACCESSES } from '../../../../../utils/hooks/useGranted';
-import Security from '../../../../../utils/Security';
-import GroupEntitySettingHiddenTypesList from '../../groups/GroupEntitySettingHiddenTypesList';
-import SettingsOrganizationEntitySettingHiddenTypesList from '../../organizations/SettingsOrganizationEntitySettingHiddenTypesList';
 import { EntitySettingsFragment_entitySetting$key } from './__generated__/EntitySettingsFragment_entitySetting.graphql';
+import EntitySettingReferences from './EntitySettingReferences';
 import { entitySettingsFragment } from './EntitySettingsFragment';
+import EntitySettingVisibility from './EntitySettingVisibility';
+import { useFormatter } from '../../../../../components/i18n';
 
 export const entitySettingPatch = graphql`
   mutation EntitySettingSettingsPatchMutation(
@@ -33,6 +26,7 @@ interface EntitySettingSettingsProps {
 
 const EntitySettingSettings = ({ entitySettingsData }: EntitySettingSettingsProps) => {
   const { t_i18n } = useFormatter();
+
   const entitySetting = useFragment(entitySettingsFragment, entitySettingsData);
   if (!entitySetting) {
     return <ErrorNotFound />;
@@ -51,130 +45,20 @@ const EntitySettingSettings = ({ entitySettingsData }: EntitySettingSettingsProp
   return (
     <Grid container={true} spacing={2}>
       <Grid item xs={6}>
-        <div>
-          <Label action={(
-            <Tooltip
-              title={!entitySetting.availableSettings.includes('platform_hidden_type')
-                ? t_i18n('This configuration is not available for this entity type')
-                : t_i18n('This configuration hides a specific entity type across the entire platform.')
-              }
-            >
-              <InformationOutline
-                fontSize="small"
-                color="primary"
-              />
-            </Tooltip>
-          )}
-          >
-            {t_i18n('Hidden in interface')}
-          </Label>
-
-          <FormGroup>
-            <FormControlLabel
-              control={(
-                <Switch
-                  disabled={
-                    !entitySetting.availableSettings.includes(
-                      'platform_hidden_type',
-                    )
-                  }
-                  checked={entitySetting.platform_hidden_type ?? false}
-                  onChange={() => handleSubmitField(
-                    'platform_hidden_type',
-                    !entitySetting.platform_hidden_type,
-                  )
-                  }
-                />
-              )}
-              label={t_i18n('Hide in the platform')}
-            />
-          </FormGroup>
-        </div>
-        <Security needs={[SETTINGS_SETACCESSES]}>
-          <>
-            <GroupEntitySettingHiddenTypesList targetType={entitySetting.target_type} />
-            <SettingsOrganizationEntitySettingHiddenTypesList targetType={entitySetting.target_type} />
-          </>
-        </Security>
+        <Card title={t_i18n('Visibility')}>
+          <EntitySettingVisibility
+            entitySetting={entitySetting}
+            handleSubmitField={handleSubmitField}
+          />
+        </Card>
       </Grid>
       <Grid item xs={6}>
-        <div>
-          <Label action={(
-            <Tooltip
-              title={!entitySetting.availableSettings.includes('platform_entity_files_ref')
-                ? t_i18n('This configuration is not available for this entity type')
-                : t_i18n('This configuration enables an entity to automatically construct an external reference from the uploaded file.')
-              }
-            >
-              <InformationOutline
-                fontSize="small"
-                color="primary"
-              />
-            </Tooltip>
-          )}
-          >
-            {t_i18n('Automatic references at file upload')}
-          </Label>
-
-          <FormGroup>
-            <FormControlLabel
-              control={(
-                <Switch
-                  disabled={
-                    !entitySetting.availableSettings.includes(
-                      'platform_entity_files_ref',
-                    )
-                  }
-                  checked={entitySetting.platform_entity_files_ref ?? false}
-                  onChange={() => handleSubmitField(
-                    'platform_entity_files_ref',
-                    !entitySetting.platform_entity_files_ref,
-                  )
-                  }
-                />
-              )}
-              label={t_i18n('Create external reference at upload')}
-            />
-          </FormGroup>
-        </div>
-        <Box sx={{ marginTop: 2 }}>
-          <Label action={(
-            <Tooltip
-              title={!entitySetting.availableSettings.includes('enforce_reference')
-                ? t_i18n('This configuration is not available for this entity type')
-                : t_i18n('This configuration enables the requirement of a reference message on an entity creation or modification.')
-              }
-            >
-              <InformationOutline
-                fontSize="small"
-                color="primary"
-              />
-            </Tooltip>
-          )}
-          >
-            {t_i18n('Enforce references')}
-          </Label>
-          <FormGroup>
-            <FormControlLabel
-              control={(
-                <Switch
-                  disabled={
-                    !entitySetting.availableSettings.includes(
-                      'enforce_reference',
-                    )
-                  }
-                  checked={entitySetting.enforce_reference ?? false}
-                  onChange={() => handleSubmitField(
-                    'enforce_reference',
-                    !entitySetting.enforce_reference,
-                  )
-                  }
-                />
-              )}
-              label={t_i18n('Enforce references')}
-            />
-          </FormGroup>
-        </Box>
+        <Card title={t_i18n('References')}>
+          <EntitySettingReferences
+            entitySetting={entitySetting}
+            handleSubmitField={handleSubmitField}
+          />
+        </Card>
       </Grid>
     </Grid>
   );
